fix: change the field surnameNamePatronymic to firstName and lastName, add the fields username and password

+++ src/main/java/usa/badratdinova/bookexchange/controllers/PeopleController.java
@@ -75,7 +75,7 @@ public class PeopleController {
     }
 
     @GetMapping("{id}")
-    public String show(@PathVariable("id") int id, Model model) {
+    public String show(@PathVariable("id") Long id, Model model) {
         model.addAttribute("person", peopleService.findOne(id));
         model.addAttribute("books", peopleService.findBooksByPersonId(id));
         return "people/show";
@@ -99,14 +99,14 @@ public class PeopleController {
     }
 
     @GetMapping("/{id}/edit")
-    public String edit(Model model, @PathVariable("id") int id) {
+    public String edit(Model model, @PathVariable("id") Long id) {
         model.addAttribute("person", peopleService.findOne(id));
         return "people/edit";
     }
 
     @PatchMapping("/{id}")
     public String update(@ModelAttribute("person") @Valid Person person, BindingResult bindingResult,
-                         @PathVariable("id") int id) {
+                         @PathVariable("id") Long id) {
         personValidator.validate(person, bindingResult);
 
         if (bindingResult.hasErrors()) {
@@ -117,7 +117,7 @@ public class PeopleController {
     }
 
     @DeleteMapping("{id}")
-    public String delete(@PathVariable("id") int id) {
+    public String delete(@PathVariable("id") Long id) {
         peopleService.delete(id);
         return "redirect:/people";
     }
